Update route redirects (#172)
Reworked routing redirects for login, consent, task, and invitation pages in CarpStudyApp. Now the routing considers if a user is authenticated and properly handles redirects to the consent or tasks pages. Furthermore, if a study ID is not present, redirections to the invitation page have been implemented. If a user is unauthenticated, they'll be redirected to the login page.